requirements.txt
Django==5.2.1
djangorestframework==3.16.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
drf-yasg==1.21.10